arreglando eliminacion de img  de  producto

+++ src/controllers/productController.js
@@ -253,17 +253,21 @@ const productController = {
                     
                     imgSecArray.forEach( async imgSec=>{
                         log('ID Img a Eliminar',imgSec);
-                        let imgS = await db.Image_product.findByPk(imgSec)
-                        fs.unlink(path.join(__dirname,`../../public/images/products/${imgS.urlName}`),(err=>{
-                            if(err) {
-                                log('Eliminacion de imgs Sec:',err);
-                            }else{ 
-                                console.log("\nDeleted file: example_file.txt");
-                            }
-                        }))
-                        db.Image_product.destroy({where:{
-                            id:imgSec
-                        }})
+                        if(imgSec !=undefined)
+                        {
+                            let imgS = await db.Image_product.findByPk(imgSec)
+                            fs.unlink(path.join(__dirname,`../../public/images/products/${imgS.urlName}`),(err=>{
+                                if(err) {
+                                    log('Eliminacion de imgs Sec:',err);
+                                }else{ 
+                                    console.log("\nDeleted file: example_file.txt");
+                                }
+                            }))
+                            db.Image_product.destroy({where:{
+                                id:imgSec
+                            }})
+                        }
+                        
                     })
                 }
